requirements-load-dataset.txt
absl-py==1.3.0
addict==2.4.0
--extra-index-url https://shared:HF6w0RbukY@packages.allegro.ai/repository/allegroai/simple
allegroai==3.6.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==20.3.0
boto3>=1.26.27
botocore==1.29.32
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
clearml==1.8.3
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
distlib==0.3.6
exceptiongroup==1.0.4
filelock==3.8.2
flatbuffers==22.12.6
fonttools==4.38.0
furl==2.1.3
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
humanfriendly==9.2
hydra-core==1.3.0
idna==3.4
iniconfig==1.1.1
jmespath==1.0.1
joblib==1.2.0
jsonschema==3.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
luqum==0.11.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mmcv==1.7.0
model-index==0.1.11
numpy==1.22.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
ordered-set==4.1.0
orderedmultidict==1.0.1
packaging==21.3
pandas==1.5.2
pathlib2==2.3.7.post1
Pillow==9.2.0
platformdirs==2.6.0
pluggy==1.0.0
ply==3.11
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==2.4.7
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.0+cu117
torchvision==0.14.0+cu117
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
Werkzeug==2.2.2
wrapt==1.14.1
yapf==0.32.0